% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{age_sex_us}
\alias{age_sex_us}
\title{age and sex}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3220 rows and 113 columns.
}
\source{
\href{https://data.census.gov/cedsci/table?q=United\%20States}{American Community Survey tables}
}
\usage{
data(age_sex_us)
}
\value{
a tibble
}
\description{
Sex and age composition of the county population from the American Community Survey of 2018
}
\keyword{datasets}
