% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{bweight_it}
\alias{bweight_it}
\title{body-mass index}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 5 columns.
}
\source{
\href{http://dati.istat.it/?lang=en}{ISTAT}
}
\usage{
data(bweight_it)
}
\description{
Body mass index in regions of Italy, in the general population.
Data were collected in 2018 and indicate absolute number of people underweight, normalweight, overweight or obese.
}
\details{
\href{http://siqual.istat.it/SIQual/visualizza.do?id=0058000&refresh=true&language=EN}{methodology}
}
\keyword{datasets}
