% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{cancer_it}
\alias{cancer_it}
\title{cancer patients}
\format{
An object of class \code{data.frame} with 21 rows and 10 columns.
}
\source{
\href{http://www.registri-tumori.it/PDF/AIOM2016/I_numeri_del_cancro_2016.pdf}{Istituto Superiore Sanita'}
}
\usage{
data(cancer_it)
}
\value{
a tibble
}
\description{
Number of cancer patients in each region by type.
Data were collected in 2016 and indicate absolute number of people diagnosed with cancer. Data for P.A. Trento and
P.A. Bolzano are missing (but we have Trentino Alto Adige)
}
\keyword{datasets}
