% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{chronic_it}
\alias{chronic_it}
\title{chronic conditions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 14 columns.
}
\source{
\href{http://dati.istat.it/?lang=en}{ISTAT}
}
\usage{
data(chronic_it)
}
\value{
a tibble
}
\description{
Number of people suffering of chronic conditions by region and type.
Data were collected in 2018 and indicate absolute number of people.
}
\details{
\href{http://siqual.istat.it/SIQual/visualizza.do?id=0058000&refresh=true&language=EN}{methodology}
}
\keyword{datasets}
