% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{firstaid_it}
\alias{firstaid_it}
\title{first aid}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.
}
\source{
\href{http://dati.istat.it/?lang=en}{ISTAT}
}
\usage{
data(firstaid_it)
}
\value{
a tibble
}
\description{
Number of people using first aid or medical guard in 3 months preceding the survey. Collected in 2018
}
\details{
\href{http://siqual.istat.it/SIQual/lang.do?language=UK}{methodology}
}
\keyword{datasets}
