% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getit_functions.R
\name{getit_all}
\alias{getit_all}
\title{get COVID-19 cases and other statistics}
\source{
\href{https://github.com/pcm-dpc/COVID-19}{protezione civile}, \href{http://dati.istat.it/?lang=en}{ISTAT}
}
\usage{
getit_all()
}
\value{
a tibble with following 64 variables:
\describe{
\item{date}{date of data}
\item{state}{state}
\item{region_code}{region abbreviation}
\item{region}{full name of region}
\item{lat}{lat}
\item{long}{long}
\item{imm}{influenza vaccination coverage in the general population}
\item{imm65}{influenza vaccination coverage in people age 65 or older}
\item{cmr}{case-mortality rate for that region and that date (deaths/total_cases * 100)}
\item{total_cases}{number of COVID-19 positive cases detected}
\item{deaths}{number of deaths}
\item{tests}{number of tests performed}
\item{hospitalized_with_symptoms}{number of people hospitalized with symptoms, that day}
\item{intensive_care_unit}{number of people in intensive care units, that day}
\item{total_hospitalized}{hospitalized_with_symptoms + intensive_care_unit}
\item{home_quarantine}{number of people COVID-19 positive in home quarantine, that day}
\item{total_positives}{total currently positives: hospitalized_with_symptoms + intensive_care_unit + home_quarantine}
\item{change_positives}{change in the number of positive cases: total_positives that day - total_positives preceding day}
\item{new_positives}{number of new positive cases: total_cases that day - total_cases preceding day}
\item{recovered_released}{recovered - released from hospital}
\item{people_tested}{number of people tested}
\item{p_house}{number of people per squared meter living in the same house}
\item{pop_tot}{total population}
\item{area_km2}{household crowding index (number of components of household per square meter)}
\item{pop_km2}{density of population per squared kilometer}
\item{female_65m}{percent of females age 65 years old or more}
\item{male_65m}{percent of males age 65 years old or more}
\item{chronic_ \emph{type}}{percent of population with that chronic condistion}
\item{cancer_\emph{type}}{percent of population with that type of cancer}
\item{bweight_\emph{type}}{percent of people underweight, normalweight, overweight or obese}
\item{first_aid}{number of peple using first aid in 3 months preceding the survey}
\item{medical_guard}{number of people using medical guard in 3 months preceding the survey}
\item{bed_acute}{inpatient hospital beds per 1000 people in acure care}
\item{bed_long}{inpatient hospital beds per 1000 people in long care}
\item{bed_rehab}{inpatient hospital beds per 1000 people in rehabilitation}
\item{bed_tot}{inpatient hospital beds per 1000 people, total}
\item{netinc}{median net annual households income, in euros}
\item{pm2.5}{emission of pm2.5 in tons per region, 2017}
}
}
\description{
extracts and translates time series form the git repository of the \href{https://github.com/pcm-dpc/COVID-19}{protezione civile} and
combines them with other statistics related to italian population
}
\details{
Data regarding COVID-19 comes form the repository of the \href{https://github.com/pcm-dpc/COVID-19}{protezione civile} and it is updated daily.
Age and sex of the population (2019),  first aid and medical guard visits (2018), smoking status (2018),  prevalence of chronic conditions (2018), annual-household income (2017)
household crowding index (2018) and body-mass index were dataset collect by  \href{http://dati.istat.it/?lang=en}{ISTAT}.
Prevalence of types of cancer patients (2016), influenza-vaccination coverage (2019) and the number of hospital beds per 1000 people (2017) were obtained from
\href{http://www.dati.salute.gov.it/}{Ministero della Salute}. Note that cancer patients prevalence was calculated using
region population esitmates of 2019. Data of particulate 2.5 (2017) comes from the
\href{https://annuario.isprambiente.it/pon/basic/14}{Istituto Superiore Per La protexione Ambientale}.
}
\seealso{
for details regarding  the methodology of specific datasets check  \code{\link{bweight_it}}, \code{\link{cancer_it}},
\code{\link{chronic_it}}, \code{\link{dem_it}}, \code{\link{firstaid_it}}, \code{\link{fl_it}}, \code{\link{fl65_it}},\code{\link{fl_it}},
\code{\link{hospbed_it}}, \code{\link{house_it}}, \code{\link{pm2.5_it}}
}
