% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getus_functions.R
\name{getus_all}
\alias{getus_all}
\title{get COVID-19 and other metrics}
\source{
\href{https://www.cms.gov/About-CMS/Agency-Information/OMH/Downloads/Mapping-Technical-Documentation.pdf}{Center for Medicare and Medicaid Services},
\href{https://hifld-geoplatform.opendata.arcgis.com/datasets/hospitals/data?page=18}{Homeland Infrastructure Foundation-Level Data},
\href{https://data.census.gov/cedsci/table?q=United\%20States}{American Community Survey tables},
\href{https://data.cms.gov/mapping-medicare-disparities}{Mapping Medicare Disparities},
\href{https://github.com/COVIDExposureIndices/COVIDExposureIndices}{COVIDExposureIndices},
\href{http://fizz.phys.dal.ca/~atmos/martin/?page_id=140#V4.NA.02.MAPLE}{Atmoshpheric Composition Analysis Group}
}
\usage{
getus_all(repo = "jhu")
}
\arguments{
\item{repo}{repository of COVID-19 data, one of \code{c("nyt", "jhu")}}
}
\value{
A dataframe. Data regarding the household composition, population sex, age, race, ancestry and poverty levels,
were scraped from the 2018 American Community Survey (ACS). Poverty was defined at the family level and not the household level in
the ACS. Medical conditions, tobacco use, cancer and, data relative to the number of medical and emergency visits
was obtained from the 2017 Mapping Medicare Disparities. From relative documentation listed in the source: "Prevalence rates are calculated
by searching for certain diagnosis codes in \strong{Medicare beneficiaries’ claims}. The admission rate by admission type is the frequency of
a specific type of inpatient admission per 1,000 inpatient admissions in a year."
The number of hospital beds per county was calculated from data of the2020 Homeland Infrastructure Foundation.
Emissions of particulate 2.5 in micro g/m3 (2000-2016) and seasonal temperature (2000-2016) were reported by \href{http://fizz.phys.dal.ca/~atmos/martin/?page_id=140#V4.NA.02.MAPLE}{Atmoshpheric Composition Analysis Group} and
aggregate by \href{https://github.com/wxwx1993/PM_COVID/blob/master/additional_preprocessing_code/download_pm25_values.md}{Ista Zahn and Ben Sabath}. \cr
The following list of variables is divided in sections \emph{COVID-19 VARS, HOUSEHOLDS MARITAL STATUS AND COMPOSITION, HOUSEHOLDS EDUCATION DEGREES,
ANCESTRY, COMPUTER OR INTERNET, POPULATION AND SEX, POPULATION AND RACE, MEDICAL AND VACCINES, POVERTY, ACTIVITY, POLLUTIONS AND TEMPERATURE, STATE LEVEL TESTS AND HOSPITALIZATIONS}. \cr
\strong{Note that data on test and hospitalizations are at the state level!}
\describe{
\item{date}{formatted \verb{ISO 8601}}
\item{county}{county}
\item{state}{state}
\item{fips}{federal information processing standard, a unique numeric identifier of a county.
Unknown fips are coded as 00000. \strong{Note that in the nyt repository a lot of deaths
and confirmed cases are no categorized
at the county level}}
\item{urban}{urban or rural (see \href{https://www.census.gov/programs-surveys/geography/guidance/geo-areas/urban-rural.html}{cenus})}
\item{\strong{COVID-19 VARS}}{---------------}
\item{cases}{confirmed COVID-19 cases (cumulates with date)}
\item{deaths}{number of deaths attributed to COVID-19}
\item{cmr}{case-mortality rate (deaths / confirmed cases * 100)}
\item{\strong{HOUSEHOLDS MARITAL STATUS AND COMPOSITION}}{---------------}
\item{total_households}{total number of households (occupy a housing unit) in that county. People not living in households are classified as living in group quarters}
\item{perc_families}{percent of households that are defined as family. A family consists of a householder and one or more other people living in the same household who are related to the householder by birth, marriage, or adoption}
\item{perc_families_18childereen}{percent families with at least a child <= 18 years old}
\item{perc_married_couples}{percent families consisting of married couples}
\item{perc_married_couples_u18ychildreen}{percent families consisting of married couples at least a child 18 years old or less}
\item{perc_families_only_male}{percent of family with a male householder and no spouse of householder present}
\item{perc_families_only_male_18ychildreen}{percent families with male householder and no spouse of householder present and with at least a child under 18 years old}
\item{perc_families_only_female}{percent families with female householder}
\item{perc_families_only_female_18ychildreen}{percent families with female householder with at least a child under 18 years old}
\item{perc_non_families}{percent of non-family households. A family consists of a householder and one or more other people living in the same household who are related to the householder by birth, marriage, or adoption}
\item{perc_non_families_alone}{percent of non-family households with householder living alone}
\item{perc_non_families_alone65y}{percent of non-family households with householder living alone, age 65 years and older}
\item{perc_non_families_u18y}{percent of non-family households with one or more people under 18 years}
\item{perc_non_families_65y}{percent of non-family households with with one or more people 65 years and older}
\item{total_relationship_in_households}{total number of people that responded to the question regarding relationship}
\item{perc_relationship_spouse}{households including person married to and living with the householder}
\item{perc_relationship_child}{households including a son or daughter by birth, a stepchild, or adopted child of the householder}
\item{perc_relationship_other_relatives}{percent households including other relatives}
\item{perc_relationship_other_nonrelatives}{percent households including foster children, not related to the householder by birth, marriage, or adoption}
\item{perc_relationship_other_unmaried_part}{percent households containing members other than a “married-couple household” that includes a householder and an “unmarried partner.” }
\item{total_marital_status_male}{total males that responded to the marital status question}
\item{perc_marital_status_male_nevermaried}{percent males never married}
\item{perc_marital_status_male_maried}{percent males married}
\item{perc_marital_status_male_separated}{percent of males separate}
\item{perc_marital_status_male_}{percent of males widowed}
\item{perc_marital_status_male_divorced}{percent of males divorced}
\item{perc_marital_status_female_nevermaried}{perent of female never married}
\item{perc_marital_status_female_maried}{perent of female married}
\item{perc_marital_status_female_separated}{perent of female separated}
\item{perc_marital_status_female_widowed}{perent of female widowed}
\item{perc_marital_status_female_divorced}{perent of female divorced}
\item{\strong{HOUSEHOLDS EDUCATION DEGREES}}{---------------}
\item{total_enrolled_school}{total people enrolled in school}
\item{perc_enrolled_preschool}{percent in preschool}
\item{perc_enrolled_kindergarden}{percent in kindergarden}
\item{perc_enrolled_elementary}{percent in elementary}
\item{perc_enrolled_highschool}{percent in highschool}
\item{perc_enrolled_college}{percent college}
\item{total_edu}{total number of people 25 years old or more that responded to the question regarding education (?)}
\item{perc_edu_9grade}{percent that went up to 9th grade}
\item{perc_edu_nodiploma}{percent that went up to 9th grade}
\item{perc_edu_highschool}{percent with highschool}
\item{perc_edu_somecollege}{percent with some college}
\item{perc_edu_associate}{percent that obtaibed an associate degree}
\item{perc_edu_bachelor}{percent with bachelor}
\item{perc_edu_gradprofess}{percent that graduated or with a professional degree}
\item{perc_edu_bachelor_higher}{percent with bachelor or higher}
\item{\strong{ANCESTRY}}{---------------}
\item{total_ancestry}{total population}
\item{perc_ \emph{anchestry}}{percent estimated specific ancestry (27)}
\item{\strong{COMPUTER OR INTERNET}}{---------------}
\item{total_withcomputer}{total that own or use a computer}
\item{perc_withcomputer}{percent that owns or use computer}
\item{perc_withinternet}{percet that has acces to internet}
\item{\strong{POPULATION AND SEX}}{---------------}
\item{total_pop}{total population}
\item{total_male}{total male}
\item{total_female}{total female}
\item{total_ \emph{age_sex}}{total population by age bin and sex}
\item{perc_ \emph{age_sex}}{percent population by age bin and sex}
\item{median_age}{median age in years}
\item{median_age_male}{median age in years of males}
\item{median_age_female}{median age in years of females}
\item{sex_ratio}{males per 100 females}
\item{age_dependency}{the age dependency ratio is derived by dividing the combined under
18  and 65-more year populations by the 18-to-64 population and multiplying the result by 100}
\item{old_age_dependency}{the old-age dependency ratio is derived by dividing the population 65 years and over by the 18-to-64 population and multiplying by 100}
\item{child_dependency}{the child dependency ratio is calculated dividing the population under 18 years by the 18-to-64 population, and multiplying the result by 100}
\item{\strong{POPULATION AND RACE}}{---------------}
\item{total_white}{total white}
\item{total_black}{total black or afroamerican}
\item{total_native}{total native}
\item{total_asian}{total asian}
\item{total_pacific_islander}{total hawaian and pacific islander}
\item{total_other_race}{other races}
\item{total_two_more_races}{two or more races}
\item{total_latino}{total hispanic or latino}
\item{\strong{MEDICAL AND VACCINES}}{---------------}
\item{perc_imm65}{percentage of fee-for-service (FFS) Medicare enrollees that had an annual flu vaccination.}
\item{total_beds}{total number of hospital beds}
\item{perc_at_least_1_chronic}{percent medicare with at least a chronic condition}
\item{perc_acute_myocardial_infarction}{percent medicare with acute myocardial infarction}
\item{perc_alzheimer_dementia}{percent medicare with Alzheimer’s Disease, Related Disorders, or Senile Dementia}
\item{perc_asthma}{percent medicare with asthma}
\item{perc_atrial_fibrillation}{percent medicare with Atrial Fibrillation}
\item{perc_cancer_breast}{percent medicare with Breast Cancer}
\item{perc_cancer_colorectal}{percent medicare with Colorectal Cancer}
\item{perc_cancer_lung}{percent medicare withLung Cancer}
\item{perc_cancer_all}{percent medicare with Cancer (breast, colorectal, lung, and/or prostate)}
\item{perc_ch_obstructive_pulm}{percent medicare with Chronic Obstructive Pulmonary Disease (COPD)}
\item{perc_chronic_kidney_disease}{percent medicare with Chronic Kidney Disease}
\item{perc_depression}{percent medicare with Depression}
\item{perc_diabetes}{percent  medicare beneficiaries with Diabetes}
\item{perc_hypertension}{percent  medicare beneficiaries with Hypertension}
\item{perc_ischemic_heart_disease}{percent  medicare beneficiaries with Ischemic Heart Disease}
\item{perc_obesity}{percent  medicare beneficiaries with Obesity}
\item{perc_osteoporosis}{percent  medicare beneficiaries with Osteoporosis}
\item{perc_rheumatoid_arthritis}{percent  medicare beneficiaries with Rheumatoid Arthritis}
\item{perc_schizophrenia_psychotic_dis}{percent  medicare beneficiaries with Schizophrenia/Other Psychotic Disorders}
\item{perc_stroke}{percent  medicare beneficiaries with Stroke Transient Ischemic Attack }
\item{perc_tobacco_use}{}
\item{urgent_admission}{urgent care admission rate}
\item{annual_wellness_visit}{number of annual wellness visits}
\item{elective_admission}{elective admission rate}
\item{emergent_admission}{ER admission rate}
\item{other_admission}{other admission rates}
\item{perc_pneumococcal_vaccine}{percent pneumococcal vaccine }
\item{\strong{POVERTY}}{---------------}
\item{total_poverty_determination}{number of people evaluated for poverty}
\item{total_poverty}{total people that met the definition of below poverty level}
\item{perc_poverty}{percent people that met the definition of below poverty level}
\item{total_determination \emph{age}}{total people evaluated in that age bin}
\item{total_poverty \emph{age}}{total people that met the definition of below poverty level in that age bin}
\item{perc_poverty \emph{age}}{percent people that met the definition of below poverty level in that age bin}
\item{total_determination \emph{sex}}{total people evaluated for poverty in that sex}
\item{total_poverty \emph{sex}}{total people that met the definition of below poverty level in that sex}
\item{perc_poverty \emph{sex}}{perc people that met the definition of below poverty level in that sex}
\item{total_determination \emph{race}}{total people evaluated for poverty in that race}
\item{total_poverty \emph{race}}{total people that met the definition of below poverty level in that race}
\item{perc_poverty \emph{race}}{perc people that met the definition of below poverty level in that race}
\item{median_income)}{median household income}
\item{\strong{ACTIVITY}}{---------------}
\item{dex_a}{activity index}
\item{\strong{POLLUTIONS AND TEMPERATURE}}{---------------}
\item{pm2.5}{pm2.5 in  micro g per m3}
\item{summer_temp}{mean temperature in summer, \%}
\item{summer_hum}{mean humity in summer, mixing ratio}
\item{winter_temp}{mean temperature in winter, K}
\item{winter_hum}{mean humity in winter, \%}
\item{\strong{STATE LEVEL TESTS AND HOSPITALIZATIONS}}{---------------}
\item{positive}{total cumulative positive test results}
\item{negative}{total cumulative negative test results}
\item{pending}{tests that have been submitted to a lab but no results have been reported yet}
\item{hospitalized_curr}{current people hospitalized}
\item{hospitalized_cumul}{cumulative people hospitalized}
\item{icu_curr}{current people in ICU}
\item{icu_cumul}{cumulative people in ICU}
\item{ventilator_curr}{current people using ventilator}
\item{ventilator_cumul}{cumulativepeople using ventilator}
\item{recovered}{total people recoverd}
\item{death_increase}{increase in deaths from day before}
\item{hospitalized_increase}{increase in hospitalization from day before}
\item{negative_increase}{increase in negative results from day before}
\item{positive_increase}{increase in positive results from day before}
\item{total_test_increase}{increase from the day before}
}
}
\description{
extracts/joins COVID-19 info with other demographic metrics at the county level and tests and hospitalizations from
\href{https://covidtracking.com/api}{the COVID Tracking Project}
}
\details{
For details regarding some specific datasets refer to: \href{https://www2.census.gov/programs-surveys/acs/tech_docs/subject_definitions/2018_ACSSubjectDefinitions.pdf?#}{Subject Definitions of the American Community Survey},
\href{https://www.cms.gov/About-CMS/Agency-Information/OMH/Downloads/Mapping-Technical-Documentation.pdf}{Medicare and Medicaid Medical Services Technical Documentation},
\href{https://github.com/COVIDExposureIndices/COVIDExposureIndices}{COVIDExposureIndices}
}
\seealso{
\code{\link{getus_covid}},\code{\link{getus_tests}}, \code{\link{getus_dex}},
}
