% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getus_functions.R
\name{getus_covid}
\alias{getus_covid}
\title{get COVID-19}
\usage{
getus_covid(repo = "jhu")
}
\arguments{
\item{repo}{repository of COVID-19 data, one of \code{c("nyt", "jhu")}}
}
\value{
a dataframe
}
\description{
extracts time series from the git repository of the  \href{ https://github.com/nytimes/covid-19-data }{NYT} or of the
\href{ https://github.com/CSSEGISandData }{JHU}
}
\details{
\code{cases} represents the number of confirmed cases, while \code{cmr} the case-mortality rate (deaths / confirmed_case * 100).
A good description of pitfalls and caveats associated with the use of case-mortality rate metric has been made on
\href{ https://ourworldindata.org/covid-mortality-risk }{Our World in Data}.
}
\examples{
dat  <- getus_covid(repo = "jhu")
}
