% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getus_functions.R
\name{getus_covid_nyt}
\alias{getus_covid_nyt}
\title{get COVID-19 from NYT}
\usage{
getus_covid_nyt()
}
\value{
a dataframe
}
\description{
extracts time series from the git repository of the \href{ https://github.com/nytimes/covid-19-data }{NYT}
}
\details{
\code{cases} represents the number of confirmed cases, while \code{cmr} the case-mortality rate (deaths / confirmed_case * 100).
A good description of pitfalls and caveats associated with the use of case-mortality rate metric has been made on
\href{ https://ourworldindata.org/covid-mortality-risk }{Our World in Data}.
}
\keyword{internal}
