% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getus_functions.R
\name{getus_dex}
\alias{getus_dex}
\title{get device-exposure indexes (DEX)}
\usage{
getus_dex()
}
\value{
a dataframe
}
\description{
extracts DEX from the git repository of the
\href{https://github.com/COVIDExposureIndices/COVIDExposureIndices}{COVID-19 exposure indeces}
}
\details{
main metric is \code{dex_a}. In the \href{https://github.com/COVIDExposureIndices/COVIDExposureIndices}{repository}, they
explains: \cite{In the context of the ongoing pandemic, the DEX measure may be biased if devices sheltering-in-place
are not in the sample due to lack of movement. We report adjusted DEX values to help address this selection bias.
DEX-adjusted is computed assuming that the number of devices has not declined since the early-2020 peak
and that unobserved devices did not visit any commercial venues.} Datataset is updated by the mantainers every weekend.
}
