% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getus_functions.R
\name{getus_tests}
\alias{getus_tests}
\title{get number of tests and  hospitalizations}
\usage{
getus_tests()
}
\value{
a dataframe with 15 variables
}
\description{
extracts information on tests, hospitalizations and other metrics at the \strong{State level} maintained by the
\href{https://covidtracking.com/api}{the COVID Tracking Project}
}
\details{
a description of the variable can be found in the \href{https://covidtracking.com/api}{the COVID Tracking Project} and when possible
was used verbatim for the description below
\describe{
\item{date}{in \verb{ISO 8601} format}
\item{state}{state name}
\item{abbr}{abbreviation}
\item{positive}{total cumulative positive test results}
\item{negative}{total cumulative negative test results}
\item{pending}{tests that have been submitted to a lab but no results have been reported yet}
\item{hospitalized_curr}{current people hospitalized}
\item{hospitalized_cumul}{cumulative people hospitalized}
\item{icu_curr}{current people in ICU}
\item{icu_cumul}{cumulative people in ICU}
\item{ventilator_curr}{current people using ventilator}
\item{ventilator_cumul}{cumulative people using ventilator}
\item{recovered}{total people recoverd}
\item{hash}{unique ID changed every time the data updates}
\item{date_checked}{date of the time we last visited their website}
\item{death}{number of deaths}
\item{death_increase}{increase in deaths from day before}
\item{hospitalized_increase}{increase in hospitalization from day before}
\item{negative_increase}{increase in negative results from day before}
\item{positive_increase}{increase in positive results from day before}
\item{total_test_increase}{increase from the day before}
}
Other details regarding the score system used are reported in the \href{https://covidtracking.com/about-data}{maintainers webpage}.\cr
\strong{Note for the use of some of some this variables by covidtracking authors:} \cr
\emph{States are currently reporting two fundamentally unlike statistics: current hospital/ICU admissions and cumulative hospitalizations/ICU admissions.
Across the country, this reporting is also sparse.
In short: it is impossible to assemble anything resembling the real statistics for hospitalizations,
ICU admissions, or ventilator usage across the United States. As a result, we will no longer provide
national-level summary hospitalizations, ICU admissions, or ventilator usage statistics on our site.}
}
