% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_us.R
\docType{data}
\name{hospbeds_us}
\alias{hospbeds_us}
\title{hospital beds}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2545 rows and 3 columns.
}
\source{
\href{https://hifld-geoplatform.opendata.arcgis.com/datasets/hospitals/data?page=18}{Homeland Infrastructure Foundation-Level Data}
}
\usage{
data(hospbeds_us)
}
\value{
a tibble
}
\description{
beds of each hospital by county.
}
\keyword{datasets}
