% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{house_it}
\alias{house_it}
\title{housing crowding}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 105 rows and 3 columns.
}
\source{
\href{http://dati.istat.it/?lang=en}{ISTAT}
}
\usage{
data(house_it)
}
\value{
a tibble in which \code{phouse} is number of components of household per square meter
}
\description{
Household crowding index from 2014 to 2018 in each region
}
\details{
\href{http://siqual.istat.it/SIQual/visualizza.do?id=5000170&refresh=true&language=EN}{methodology}
}
\keyword{datasets}
