% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{it_dem}
\alias{it_dem}
\title{Percent of population by region, sex and age. Data were collected in 2019 and indicate absolute number of people. Long format,}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4242 rows and 9 columns.
}
\source{
\href{http://dati.istat.it/?lang=en}{ISTAT}
}
\usage{
data(it_dem)
}
\value{
a tibble
}
\description{
Percent of population by region, sex and age. Data were collected in 2019 and indicate absolute number of people. Long format,
}
\details{
\href{http://siqual.istat.it/SIQual/visualizza.do?id=0019900&refresh=true&language=EN}{methodology}
The Istituto Superiore Sanita' provides biweekly info regarding the mortality in different age groups fro patients positive for COVID-19 in this
\href{https://www.epicentro.iss.it/coronavirus/sars-cov-2-decessi-italia}{link}
}
\keyword{datasets}
