% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{it_fl}
\alias{it_fl}
\title{influenza vaccination coverage, general population, time series}
\format{
An object of class \code{data.frame} with 21 rows and 21 columns.
}
\source{
\href{http://www.salute.gov.it/imgs/C_17_tavole_19_allegati_iitemAllegati_0_fileAllegati_itemFile_3_file.pdf}{Ministero della Salute}
}
\usage{
data(it_fl)
}
\description{
Influenza vaccination coverage in Italy in the \strong{general population} from 1999 to 2019. Data are percent of region population
}
\keyword{datasets}
