% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{it_fl65}
\alias{it_fl65}
\title{influenza vaccination coverage 2019}
\format{
An object of class \code{data.frame} with 22 rows and 21 columns.
}
\source{
\href{http://www.salute.gov.it/imgs/C_17_tavole_19_allegati_iitemAllegati_0_fileAllegati_itemFile_3_file.pdf}{Ministero della Salute}
}
\usage{
data(it_fl65)
}
\value{
a tibble with following columns:
\describe{
\item{\strong{region}}{region}
\item{\strong{perc_imm65}}{percent of population age 65 or more that received influenza vaccination}
\item{\strong{perc}}{percent of general population that received influenza vaccination}
}
}
\description{
Influenza vaccination coverage in Italy for 2018-2019 season for population age 65 or more from 1999 to 2019. Data are percent of region population
}
\keyword{datasets}
