% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{it_hospbed}
\alias{it_hospbed}
\title{hospital beds}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 5 columns.
}
\source{
\href{http://www.dati.salute.gov.it/}{Ministero della Salute}
}
\usage{
data(it_hospbed)
}
\value{
a tibble in wide format in which \code{bed_acute}, \code{bed_long}, \code{bed_rehab}, \code{bed_tot} refers to acute care, long term care,
rehabilitation and total beds, respectivelly
}
\description{
Inpatient hospital beds per 1000 people. Collected in 2017
}
\details{
\href{http://www.dati.salute.gov.it/dati/dettaglioDataset.jsp?menu=dati&idPag=18}{methodology}
}
\keyword{datasets}
