% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{it_netinc}
\alias{it_netinc}
\title{Net income}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 2 columns.
}
\source{
\href{http://dati.istat.it/?lang=en}{ISTAT}
}
\usage{
data(it_netinc)
}
\value{
a tibble
}
\description{
Median net annual households income (including imputed rents, in euros). Collected in 2017
}
\details{
\href{http://siqual.istat.it/SIQual/visualizza.do?id=5000170&refresh=true&language=EN}{methodology}
}
\keyword{datasets}
