% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{it_pm2.5}
\alias{it_pm2.5}
\title{particulate 2.5 italy}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 2 columns.
}
\source{
\href{https://annuario.isprambiente.it/pon/basic/14}{Istituto Superiore Per La protezione Ambientale}
}
\usage{
data("it_pm2.5")
}
\value{
a tibble
}
\description{
Emission of pm2.5 in tons per region from 1990 to 2017
}
\details{
\href{https://annuario.isprambiente.it/pon/basic/14}{methodology}
}
\keyword{datasets}
