% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{it_regions}
\alias{it_regions}
\title{regions area}
\format{
An object of class \code{data.frame} with 21 rows and 2 columns.
}
\usage{
data(it_regions)
}
\value{
a tibble
}
\description{
Area in square meters of each region. Used to calculate density per region. Scraped from old good wikipedia.
}
\keyword{datasets}
