% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{it_smoking}
\alias{it_smoking}
\title{smoking status}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 4 columns.
}
\source{
\href{http://dati.istat.it/?lang=en}{ISTAT}
}
\usage{
data(it_smoking)
}
\value{
a tibble
}
\description{
Number of people age 14 years and over that self-refer as smoker, non smoker, or past smoker by region and type.
Data were collected in 2018 and are absolute number of people.
}
\details{
\href{http://siqual.istat.it/SIQual/visualizza.do?id=0058000&refresh=true&language=EN}{methodology}
}
\keyword{datasets}
