% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_us.R
\docType{data}
\name{netinc_us}
\alias{netinc_us}
\title{netinc_us}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3220 rows and 4 columns.
}
\source{
\href{http://fizz.phys.dal.ca/~atmos/martin/?page_id=140#V4.NA.02.MAPLE}{Atmoshpheric Composition Analysis Group},

\href{https://data.census.gov/cedsci/table?q=United\%20States}{American Community Survey tables}
}
\usage{
data(netinc_us)
}
\value{
a tibble
}
\description{
Median Household income, 2018
}
\details{
\href{https://www2.census.gov/programs-surveys/acs/tech_docs/subject_definitions/2018_ACSSubjectDefinitions.pdf?#}{Subject Definitions of the American Community Survey}
}
\keyword{datasets}
