% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_it.R
\docType{data}
\name{pm2.5_it}
\alias{pm2.5_it}
\title{particulate 2.5}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 2 columns.
}
\source{
\href{https://annuario.isprambiente.it/pon/basic/14}{Istituto Superiore Per La protexione Ambientale}
}
\usage{
data("pm2.5_it")
}
\value{
a tibble
}
\description{
Emission of pm2.5 in tons per region from 1990 to 2017
}
\details{
\href{http://siqual.istat.it/SIQual/visualizza.do?id=5000170&refresh=true&language=EN}{methodology}
}
\keyword{datasets}
