% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_us.R
\docType{data}
\name{pm2.5_us}
\alias{pm2.5_us}
\title{particulate 2.5}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3176 rows and 2 columns.
}
\source{
\href{http://fizz.phys.dal.ca/~atmos/martin/?page_id=140#V4.NA.02.MAPLE}{Atmoshpheric Composition Analysis Group},
\href{https://github.com/wxwx1993/PM_COVID/blob/master/Data/county_pm25.csv}{wxwk1993 processed data}
}
\usage{
data(pm2.5_us)
}
\value{
a tibble
}
\description{
Emission of pm2.5 in  micro g/m3 per county from 2000 to 2016
}
\details{
\href{https://github.com/wxwx1993/PM_COVID/blob/master/additional_preprocessing_code/download_pm25_values.md}{Ista Zahn and Ben Sabath repo}
}
\keyword{datasets}
