% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{poverty_us}
\alias{poverty_us}
\title{poverty}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3220 rows and 63 columns.
}
\source{
\href{https://data.census.gov/cedsci/table?q=United\%20States}{American Community Survey tables}
}
\usage{
data(poverty_us)
}
\value{
a tibble
}
\description{
Household living below the poverty level, divided by age and race and calculate as absolute value or percentage. American Community Survey of 2018
}
\details{
\href{https://www2.census.gov/programs-surveys/acs/tech_docs/subject_definitions/2018_ACSSubjectDefinitions.pdf?#}{Subject Definitions of the American Community Survey}
}
\keyword{datasets}
