% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_us.R
\docType{data}
\name{race_us}
\alias{race_us}
\title{race}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3220 rows and 23 columns.
}
\source{
\href{https://data.census.gov/cedsci/table?q=United\%20States}{American Community Survey tables}
}
\usage{
data(race_us)
}
\value{
a tibble
}
\description{
Estimate population of each county by race. American Community Survey of 2018
}
\details{
\href{https://www2.census.gov/programs-surveys/acs/tech_docs/subject_definitions/2018_ACSSubjectDefinitions.pdf?#}{Subject Definitions of the American Community Survey}
}
\keyword{datasets}
