% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_us.R
\docType{data}
\name{season_us}
\alias{season_us}
\title{seasonal temperature and humidity}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3233 rows and 5 columns.
}
\source{
\href{http://fizz.phys.dal.ca/~atmos/martin/?page_id=140#V4.NA.02.MAPLE}{Atmoshpheric Composition Analysis Group},
\href{https://github.com/wxwx1993/PM_COVID/blob/master/Data/county_pm25.csv}{wxwk1993 processed data}
}
\usage{
data(season_us)
}
\value{
a tibble
}
\description{
Seasonal temperature and humidity
}
\details{
\href{https://raw.githubusercontent.com/wxwx1993/PM_COVID/master/Data/temp_seasonal_county.csv}{Ista Zahn and Ben Sabath repo}
}
\keyword{datasets}
