% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_us.R
\docType{data}
\name{us_acm_househ}
\alias{us_acm_househ}
\title{household composition}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3142 rows and 82 columns.
}
\source{
\href{https://data.census.gov/cedsci/table?q=United\%20States}{American Community Survey tables}
}
\usage{
data(us_acm_househ)
}
\value{
a tibble
}
\description{
Several metrics regarding household composition from the American Community Survey of 2018
}
\details{
\href{https://www2.census.gov/programs-surveys/acs/tech_docs/subject_definitions/2018_ACSSubjectDefinitions.pdf?#}{Subject Definitions}
}
\keyword{datasets}
