% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_us.R
\docType{data}
\name{us_mmd}
\alias{us_mmd}
\title{mapping medicare disparities}
\format{
An object of class \code{data.frame} with 3238 rows and 33 columns.
}
\source{
\href{https://data.cms.gov/mapping-medicare-disparities}{Mapping Medicare Disparities}
}
\usage{
data(us_mmd)
}
\value{
a tibble
}
\description{
Prevalence of many  medical and chronic conditions, 2017. From relative documentation listed below: "Prevalence rates are calculated by searching for certain diagnosis codes in Medicare
beneficiaries’ claims. The prevalence rate of a condition for a specific sub-population
(e.g., all beneficiaries in a county) is the proportion of beneficiaries who are found to have the condition. The admission rate by admission type is the frequency of a specific type of inpatient admission
per 1,000 inpatient admissions in a year."
}
\details{
Details regarding the use of the webtool can be found in the relative
\href{https://www.cms.gov/About-CMS/Agency-Information/OMH/Downloads/Mapping-Technical-Documentation.pdf}{documentation}. It includes prevalence of
\itemize{
\item Alzheimer
\item chronic kidney
\item obesity,
\item depression
\item obstructive pulmonary
\item disease
\item arthritis
\item diabetes
\item osteoporosis
\item asthma
\item atrial
\item fibrillation
\item ischemic hearth,
\item myocardial infarction
\item hypertension
\item several type of cancer
\item emergency,  medical admissions, annual visits
\item pneumoccocal vaccine
\item tabacco use
}
}
\seealso{
\code{\link{getus_all}} for more details regarding the variables
}
\keyword{datasets}
