% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid.R
\name{get_covid}
\alias{get_covid}
\title{get COVID-19 updated cases}
\usage{
get_covid()
}
\value{
a tibble with following columns:
\describe{
\item{\strong{date}}{date of data}
\item{\strong{state}}{state}
\item{\strong{region_code}}{region abbreviation}
\item{\strong{region}}{full name of region}
\item{\strong{lat}}{lat}
\item{\strong{long}}{long}
\item{\strong{COVID19_var}}{
\itemize{
\item cmr: case-mortality rate for that region and that date (deaths/total_cases * 100)
\item hospitalized_with_symptoms: number of people hospitalized with symptoms
\item intensive_care_unit: number of people in intensive care units
\item home_quarantine: number of people COVID-19 positive in home quarantine
\item deaths: number of deaths
\item total_cases: number of COVID-19 positive cases detected
\item tests: number of tests performed
}
}
\item{\strong{value}}{value relative to COVID19_var}

}
}
\description{
extracts and translates time series form the git repository of the \href{https://github.com/pcm-dpc/COVID-19}{protezione civile}
}
